Mostly working long node ids. Not ready for merge.
Existing unit tests are all passing. The read-only versions of the
classes (the query system versions) have not been updated to read
the new format. Also new tests should be added to test between
legacy, sparse, and long formats. Ref T51939

+++ hydra-data/src/main/java/com/addthis/hydra/data/tree/ReadTreeNode.java
@@ -100,7 +100,7 @@ public class ReadTreeNode extends AbstractTreeNode implements IReadWeighable {
         throw new UnsupportedOperationException();
     }
 
-    public Integer nodeDB() {
+    public Long nodeDB() {
         return nodedb;
     }
